backend/requirements.txt

Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
